BOSS gotoField applies red hartmann correction even if blue fails
See ticket #2701. In goto_field_boss, hartmann collimate is now called
with ignoreResiduals (same as collimateBoss) so that the red correction
is always applies. collimate is also called with the new option
minBlueCorrection which makes hartmannActor output only the minimum
correction needed to the blue ring to get within the focus tolerance.

If the hartmann is in focus, gotoField proceeds normally. If it is out
of focus, the red correction is applies and then gotoField fails,
leaving the lights on.