now doing survey specfic things for hartmann in goto_field
- If APOGEE is leading with call hartmann collimate ignoreResiduals so
that the collimator is always adjusted.
- If MaNGA is leading we call hartmann collimate ignoreResiduals
minBlueCorrection. This outputs only the minimum blue ring correction
required to get the cameras into focus tolerance. If any of the cameras
is out of focus we stop the goto_field command and turn the lamps off.
- For eBOSS we do the same this but we only call hartmann collimate
ignoreResiduals (same as collimate_boss).

Also, the hartmann code for goto_field is now its own separate function
(it has grown too big).

Updated and extended tests.