Reorder steps when aborting GotoField
The idea is that, maybe, there was enough time between the TCC was stopped and when super(GotoFieldCmd, self).abort() run, which gave the command a chance to get to the Hartmanns and from there on continue the sequence. This should first of all set the cmdState.aborting state that should stop any new command in a MultiCommand, and then proceed to cancel the slew first (which should be fast),and then the BOSS exposure.